# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Strangler Studio API
  description: |
    REST API for the Strangler Studio demonstration application.
    This API manages Student Requests with a Halloween theme.
  version: 1.0.0
  contact:
    name: Strangler Studio
    
servers:
  - url: http://localhost/api
    description: Local development server (via gateway)
  - url: http://localhost:8000
    description: Direct API service access

paths:
  /requests:
    get:
      summary: List all Student Requests
      description: Returns a list of all Student Request objects in the system
      operationId: listRequests
      tags:
        - requests
      responses:
        '200':
          description: Successful response with array of Student Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentRequest'
              examples:
                halloween_requests:
                  summary: Example Halloween-themed requests
                  value:
                    - id: 1
                      student_name: "Victor Frankenstein"
                      school: "Miskatonic University"
                      status: "Possessed"
                      created_at: "2024-10-31T23:59:59Z"
                      priority: "Critical"
                      notes: "Urgent reanimation assistance required"
                    - id: 2
                      student_name: "Mina Harker"
                      school: "Transylvania Academy"
                      status: "Summoned"
                      created_at: "2024-10-30T18:30:00Z"
                      priority: "High"
                      notes: "Vampire literature research"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /requests/{id}:
    get:
      summary: Get Student Request by ID
      description: Returns a single Student Request object matching the specified ID
      operationId: getRequestById
      tags:
        - requests
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the Student Request
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Successful response with Student Request object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentRequest'
              examples:
                single_request:
                  summary: Example Student Request
                  value:
                    id: 1
                    student_name: "Victor Frankenstein"
                    school: "Miskatonic University"
                    status: "Possessed"
                    created_at: "2024-10-31T23:59:59Z"
                    priority: "Critical"
                    notes: "Urgent reanimation assistance required"
        '404':
          description: Student Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                not_found:
                  summary: Request not found
                  value:
                    detail: "Request not found"
        '422':
          description: Validation error - invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              examples:
                invalid_id:
                  summary: Invalid ID parameter
                  value:
                    detail:
                      - loc: ["path", "id"]
                        msg: "value is not a valid integer"
                        type: "type_error.integer"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    StudentRequest:
      type: object
      required:
        - id
        - student_name
        - school
        - status
        - created_at
        - priority
      properties:
        id:
          type: integer
          description: Unique identifier for the Student Request
          example: 1
          minimum: 1
        student_name:
          type: string
          description: Full name of the student
          example: "Victor Frankenstein"
          minLength: 1
          maxLength: 200
        school:
          type: string
          description: Name of the educational institution
          example: "Miskatonic University"
          minLength: 1
          maxLength: 200
        status:
          type: string
          description: Current state of the request (Halloween-themed)
          enum:
            - Possessed
            - Banished
            - Summoned
            - Pending
          example: "Possessed"
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the request was created
          example: "2024-10-31T23:59:59Z"
        priority:
          type: string
          description: Urgency level of the request
          enum:
            - Critical
            - High
            - Medium
            - Low
          example: "Critical"
        notes:
          type: string
          description: Additional information or comments about the request
          example: "Urgent reanimation assistance required"
          maxLength: 1000
      description: Represents a request submitted by a student

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Internal server error"
      description: Standard error response

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Location of the error in the request
              msg:
                type: string
                description: Human-readable error message
              type:
                type: string
                description: Error type identifier
          description: List of validation errors
      description: Validation error response with detailed error information

tags:
  - name: requests
    description: Operations related to Student Requests